+++ x/oracle/module/module_ibc.go
@@ -1,13 +1,12 @@
 package oracle
 
 import (
-	icqtypes "github.com/strangelove-ventures/async-icq/v6/types"
-
 	cerrs "cosmossdk.io/errors"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	capabilitytypes "github.com/cosmos/cosmos-sdk/x/capability/types"
+	icqtypes "github.com/cosmos/ibc-apps/modules/async-icq/v6/types"
 	channeltypes "github.com/cosmos/ibc-go/v6/modules/core/04-channel/types"
 	porttypes "github.com/cosmos/ibc-go/v6/modules/core/05-port/types"
 	host "github.com/cosmos/ibc-go/v6/modules/core/24-host"